TextField: remove redundant field layer [+]
Turns out that the TextField had 2 layers of <div class='field'>
which interfered with layout in forms. Removing it allows us
to place the textfield correctly and get array rows to fill out
the length of the row.

I had to remove ErrorLabel from TextField to achieve this, but the
error props seems enough. I have attached screenshots to show it.